Fix Android build configuration and type error
- Remove legacy android/build.gradle that conflicted with new
  settings.gradle.kts (FAIL_ON_PROJECT_REPOS setting)
- Increase Gradle JVM heap to 2GB to avoid GC thrashing
- Fix type mismatch in WormholeApplication.kt (remove .toInt())

+++ android/app/src/main/kotlin/io/sanford/wormholewilliam/WormholeApplication.kt
@@ -20,7 +20,7 @@ class WormholeApplication : Application() {
             wormholeClient.setRendezvousURL(config.rendezvousURL)
         }
         if (config.codeLength > 0) {
-            wormholeClient.setCodeLength(config.codeLength.toInt())
+            wormholeClient.setCodeLength(config.codeLength)
         }
     }
 }
